Fixed tests and cleaned up ManimWriter.kt
