+++ src/test/testFiles/python/stack.py
@@ -104,11 +104,7 @@ class Stack(DataStructure, ABC):
         empty.all.move_to(np.array([self.width_center, self.lr[1], 0]), aligned_edge=self.aligned_edge)
         self.all.add(empty.all)
         return ShowCreation(empty.all)
-    def push(self, obj, color=None, text_color=None):
-        if not color:
-            color = self.color
-        if not text_color:
-            text_color = self.text_color
+    def push(self, obj):
         animations = []
         obj.all.move_to(np.array([self.width_center, self.ul[1] - 0.1, 0]), UP)
         shrink, scale_factor = self.shrink_if_cross_border(obj.all)
@@ -118,16 +114,10 @@ class Stack(DataStructure, ABC):
         obj.all.scale(target_width / obj.all.get_width())
         animations.append([FadeIn(obj.all)])
         animations.append([ApplyMethod(obj.all.next_to, self.all, np.array([0, 0.25, 0]))])
-        animations.append([FadeToColor(obj.shape, color), FadeToColor(obj.text, text_color)])
         return animations
-    def pop(self, obj, color=None, text_color=None, fade_out=True):
-        if not color:
-            color = self.color
-        if not text_color:
-            text_color = self.text_color
+    def pop(self, obj, fade_out=True):
         self.all.remove(obj.all)
-        animation = [[FadeToColor(obj.shape, color), FadeToColor(obj.text, text_color)],
-                     [ApplyMethod(obj.all.move_to, np.array([self.width_center, self.ul[1] - 0.1, 0]), UP)]]
+        animation = [[ApplyMethod(obj.all.move_to, np.array([self.width_center, self.ul[1] - 0.1, 0]), UP)]]
         if fade_out:
             animation.append([FadeOut(obj.all)])
             enlarge, scale_factor = self.shrink(new_width=self.all.get_width(), new_height=self.all.get_height() + 0.25)
@@ -139,11 +129,7 @@ class Stack(DataStructure, ABC):
         if self.will_cross_boundary(height, "TOP"):
             return self.shrink(new_width=self.all.get_width(), new_height=self.all.get_height() + height + 0.4)
         return 0, 1
-    def push_existing(self, obj, color=None, text_color=None):
-        if not color:
-            color = self.color
-        if not text_color:
-            text_color = self.text_color
+    def push_existing(self, obj):
         animation = [[ApplyMethod(obj.all.move_to, np.array([self.width_center, self.ul[1] - 0.1, 0]), UP)]]
         enlarge, scale_factor = obj.owner.shrink(new_width=obj.owner.all.get_width(), new_height=obj.owner.all.get_height() + 0.25)
         sim_list = list()
@@ -155,7 +141,6 @@ class Stack(DataStructure, ABC):
         if len(sim_list) != 0:
             animation.append(sim_list)
         animation.append([ApplyMethod(obj.all.next_to, self.all, np.array([0, 0.25, 0]))])
-        animation.append([FadeToColor(obj.shape, color), FadeToColor(obj.text, text_color)])
         return animation
 # Object representing a stack instantiation.
 class Init_structure:
